Added ramsey solver class
