+++ OUTPUT.txt
@@ -1,6 +1,7 @@
 matrix([
-     [1,1,1,1],
-     [1,1,1,1],
-     [1,1,1,1],
-     [1,1,1,1]
+     [0,1,0,0,1],
+     [1,0,1,0,0],
+     [0,1,0,1,0],
+     [0,0,1,0,1],
+     [1,0,0,1,0]
 ]).
